patient view update
Added profile info view, added chose date

+++ app/src/main/java/com/example/monitoringsystem/RegisterActivity.java
@@ -83,12 +83,12 @@ public class RegisterActivity extends AppCompatActivity {
                     sb.append(r.nextInt(10));
                     if (checkBox.isChecked()) {
                         sb.insert(0, 'L');
-                        User user = new User(name.getText().toString(), surname.getText().toString(), Integer.parseInt(phone.getText().toString()), sb.toString(), password.getText().toString(), Double.parseDouble(pesel.getText().toString()));
+                        User user = new User(name.getText().toString(), surname.getText().toString(), Integer.parseInt(phone.getText().toString()), sb.toString(), password.getText().toString(), Long.parseLong(pesel.getText().toString()));
                         mDatabase.child("users").child(user.login).setValue(user);
                         Toast.makeText(RegisterActivity.this, "Utworzono konto twój login:!" + user.login, Toast.LENGTH_SHORT).show();
                     } else {
                         sb.insert(0, 'P');
-                        User user = new User(name.getText().toString(), surname.getText().toString(), Integer.parseInt(phone.getText().toString()), sb.toString(), password.getText().toString(), Double.parseDouble(pesel.getText().toString()));
+                        User user = new User(name.getText().toString(), surname.getText().toString(), Integer.parseInt(phone.getText().toString()), sb.toString(), password.getText().toString(), Long.parseLong(pesel.getText().toString()));
                         mDatabase.child("users").child(user.login).setValue(user);
                         for (int i = 0; i < 3; i++) {
                             Toast.makeText(RegisterActivity.this, "Utworzono konto twój login: " + user.login + " zapamiętaj go.", Toast.LENGTH_LONG).show();
